Разделил исходный код на модули
